@@ -4,8 +4,8 @@ module com.github.Snuslyk.slib {
     requires java.sql;
     requires org.hibernate.orm.core;
     requires java.naming;
-    requires java.persistence;
     requires java.desktop;
+    requires jakarta.persistence;
 
 
     opens com.github.Snuslyk.slib to javafx.fxml;
